+++ test/cli/test_commands.py
@@ -131,6 +131,7 @@ def test_schema_not_available(cli, workers, snapshot_cli):
 
 @pytest.mark.openapi_version("3.0")
 @pytest.mark.operations("success")
+@pytest.mark.snapshot_suffix(platform.python_implementation().lower())
 def test_empty_schema_file(testdir, cli, snapshot_cli):
     # When the schema file is empty
     filename = testdir.makefile(".json", schema="")
@@ -290,6 +291,7 @@ def test_status_code_conformance(cli, schema_url, workers, snapshot_cli):
 
 
 @pytest.mark.operations("headers")
+@pytest.mark.skipif(platform.python_implementation() == "PyPy", reason="aiohttp crashes on PyPy")
 def test_headers_conformance_valid(cli, schema_url):
     result = cli.run(schema_url, "-c", "response_headers_conformance", "-H", "X-Custom-Header: 42")
     assert result.exit_code == ExitCode.OK, result.stdout
@@ -1088,6 +1090,7 @@ def test_response_schema_conformance_deduplication(cli, schema_url, snapshot_cli
 
 @pytest.mark.openapi_version("3.0")
 @pytest.mark.operations("malformed_json")
+@pytest.mark.skipif(platform.python_implementation() == "PyPy", reason="PyPy behaves differently")
 def test_malformed_json_deduplication(cli, schema_url, snapshot_cli):
     # See GH-1518
     # When responses are not JSON as expected and their content differ each time
@@ -1866,6 +1869,7 @@ def always_fails(ctx, response, case):
 
 
 @pytest.mark.skipif(platform.system() == "Windows", reason="Requires extra setup on Windows")
+@pytest.mark.skipif(platform.python_implementation() == "PyPy", reason="PyPy behaves differently")
 def test_app_crash(subprocess_runner, cli, snapshot_cli):
     app = """
 import os
@@ -1899,6 +1903,7 @@ if __name__ == "__main__":
 
 
 @pytest.mark.skipif(platform.system() == "Windows", reason="Requires extra setup on Windows")
+@pytest.mark.skipif(platform.python_implementation() == "PyPy", reason="PyPy behaves differently")
 def test_partial_response(subprocess_runner, cli, snapshot_cli):
     app = """
 import os
@@ -1940,6 +1945,7 @@ if __name__ == "__main__":
 
 @pytest.mark.skipif(platform.system() == "Windows", reason="Requires extra setup on Windows")
 @pytest.mark.snapshot(replace_phase_statistic=True)
+@pytest.mark.skipif(platform.python_implementation() == "PyPy", reason="PyPy behaves differently")
 def test_stateful_crash(subprocess_runner, cli, snapshot_cli):
     app = """
 import os