+++ test/loaders/test_openapi.py
@@ -2,6 +2,7 @@
 
 import io
 import json
+import platform
 
 import pytest
 import yaml
@@ -80,7 +81,10 @@ def test_unsupported_type():
         schemathesis.openapi.from_dict({})
 
 
-JSON_ERROR = ["Expecting property name enclosed in double quotes: line 1 column 2 (char 1)"]
+if platform.python_implementation() == "PyPy":
+    JSON_ERROR = ["Key name must be string at char: line 1 column 2 (char 1)"]
+else:
+    JSON_ERROR = ["Expecting property name enclosed in double quotes: line 1 column 2 (char 1)"]
 YAML_ERROR = [
     "unacceptable character #x0080: control characters are not allowed",
     '  in "<unicode string>", position 2',